+++ apps/ui-e2e/src/component/tab-group.component.cy.ts
@@ -6,7 +6,7 @@ describe('TabContainerComponent', () => {
   it('Renders default tab group and clicks the second tab', () => {
     cy.loadStory('Angular Tab Group', 'Default')
       .get('cvi-ng-tab-group')
-      .shouldHaveClasses('ul', 'cvi-tab-group')
+      .shouldHaveClasses('div', 'cvi-tab-group__list')
       .shouldHaveClasses('[data-cy="tab-button-0"]', [
         'cvi-tab-group__tab-btn',
         'active',