+++ libs/ui/src/lib/tabs/tab-group.component.stories.ts
@@ -29,6 +29,7 @@ const Template: Story<TabGroupComponent> = (args: TabGroupComponent) => ({
     <cvi-ng-tab-group>
       <cvi-ng-tab title="Tab 1">{{ content }}</cvi-ng-tab>
       <cvi-ng-tab title="Tab 2">Second tab content</cvi-ng-tab>
+      <cvi-ng-tab title="Tab 3 with lots and lots of random text">Third tab content. File not found. I was having the most wonderful dream. Except you were there, and you were there, and you were there! And when we woke up, we had these bodies. I could if you hadn't turned on the light and shut off my stereo.</cvi-ng-tab>
     </cvi-ng-tab-group>
   `,
 });