+++ libs/ui/src/lib/tabs/tab-group.component.ts
@@ -4,11 +4,15 @@ import {
   ChangeDetectorRef,
   Component,
   ContentChildren,
+  ElementRef,
+  HostBinding,
   OnDestroy,
   QueryList,
+  ViewChildren,
 } from '@angular/core';
 import { TabComponent } from './tab.component';
 import { merge, Subscription } from 'rxjs';
+import { v4 as uuidv4 } from 'uuid';
 
 @Component({
   selector: 'cvi-ng-tab-group',
@@ -17,11 +21,18 @@ import { merge, Subscription } from 'rxjs';
 })
 export class TabGroupComponent implements AfterViewInit, OnDestroy {
   @ContentChildren(TabComponent) allTabs!: QueryList<TabComponent>;
+  @HostBinding('class') get hostClasses(): string {
+    return 'cvi-tab-group';
+  }
 
   activeIndex = 0;
+  baseId = 'tabgroup-' + uuidv4();
 
   /** @internal */
   private tabChangesSubscription = Subscription.EMPTY;
+  @ViewChildren('tabButton') tabButtons!: QueryList<
+    ElementRef<HTMLButtonElement>
+  >;
 
   constructor(private cd: ChangeDetectorRef) {}
 
@@ -37,6 +48,34 @@ export class TabGroupComponent implements AfterViewInit, OnDestroy {
 
   makeActive(index: number) {
     this.activeIndex = index;
+  }
+
+  updateButtonFocus(): void {
+    this.tabButtons.get(this.activeIndex)?.nativeElement.focus();
+  }
+
+  makeActivePrev(event: Event) {
+    event.preventDefault();
+    if (this.activeIndex > 0) {
+      this.activeIndex--;
+    } else {
+      this.activeIndex = this.allTabs.length - 1;
+    }
+    this.updateButtonFocus();
+  }
+
+  makeActiveNext(event: Event) {
+    event.preventDefault();
+    if (this.activeIndex < this.allTabs.length - 1) {
+      this.activeIndex++;
+    } else {
+      this.activeIndex = 0;
+    }
+    this.updateButtonFocus();
+  }
+
+  isTabSelected(tabIndex: number): boolean {
+    return this.activeIndex === tabIndex;
   }
 
   getActiveTabContent() {